swapped home post feed to separate activity. - less conflict with current main branch
